New Model of NewRequest
* Add new entity NewRequest, each NewRequest has Client (ManyToOne) but Clients are not aware of this
* Add repository and service classes
* Test repository and service classes
* Change enum databaseCode to Integer to avoid more compley compatibility issues

+++ src/main/java/com/felixseifert/swedisheventplanners/model/enums/EventTypeConverter.java
@@ -5,16 +5,16 @@ import javax.persistence.Converter;
 import java.util.Arrays;
 
 @Converter
-public class EventTypeConverter implements AttributeConverter<EventType, Byte> {
+public class EventTypeConverter implements AttributeConverter<EventType, Integer> {
 
     @Override
-    public Byte convertToDatabaseColumn(EventType eventType) {
+    public Integer convertToDatabaseColumn(EventType eventType) {
         if(eventType == null) return null;
         return eventType.getDatabaseCode();
     }
 
     @Override
-    public EventType convertToEntityAttribute(Byte databaseCode) {
+    public EventType convertToEntityAttribute(Integer databaseCode) {
         if(databaseCode == null) return null;
         return Arrays.stream(EventType.values()).filter(type -> type.getDatabaseCode().equals(databaseCode))
                 .findFirst().orElseThrow(IllegalArgumentException::new);
